% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_plot.R
\name{acf_plot}
\alias{acf_plot}
\title{acf_plot}
\usage{
acf_plot(tsdata)
}
\arguments{
\item{tsdata}{the time series which you are interested in and want to see the autocorrelogram}
}
\value{

}
\description{
acf_plot
}
\examples{
data = ITStock::DGS10
CMR_ts=ts(as.numeric(data[,2]),start=c(1980,11),freq=12)
acf_plot(CMR_ts)
}
