% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMC_get_listing.R
\name{CMC_get_listing}
\alias{CMC_get_listing}
\title{CMC_get_listing}
\usage{
CMC_get_listing(API_key, start = 1, fiat = "USD", limit = 5000)
}
\arguments{
\item{API_key}{API key created through the official site of Coin Market Cap}

\item{start}{The starting date of the data, default = 1}

\item{fiat}{the currency in which the value of the cryptocurrency is presented. default is in USD}

\item{limit}{maximum number of entries to acquire, default is 5000}
}
\value{
a dataframe of the current cryptocurrency market info of interest
}
\description{
CMC_get_listing
}
\examples{
CMC_get_listing(API_key = "496cdd8f-3b52-4296-b467-fc6617012118")
}
