% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Nasdaq_100.R
\name{get_Nasdaq_100}
\alias{get_Nasdaq_100}
\title{get_Nasdaq_100}
\usage{
get_Nasdaq_100(API_key)
}
\arguments{
\item{API_key}{The API key for Quandl.com}
}
\value{
returns financial market data in a dataframe
}
\description{
get_Nasdaq_100
}
\examples{
API_key = "rw4g3B3zCDzzb-yyrKeQ"
get_Nasdaq_100(API_key)
}
