% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE}
\alias{MLE}
\title{MLE}
\usage{
MLE(likelihood_func, lower = -10, upper = 10, units = 1000)
}
\arguments{
\item{likelihood_func}{the likelihood function}

\item{lower}{the lower bound of the interval where MLE should exist}

\item{upper}{the upper bound of the interval where MLE should exist}

\item{units}{the number of equal spaces between the upper and lower bound}
}
\value{

}
\description{
MLE
}
\examples{
f = (exp(-(theta-1)^2)/2)+3*exp(-((theta-2)^2)/2)
MLE(f)
}
