% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_plot.R
\name{stock_plot}
\alias{stock_plot}
\title{stock_plot}
\usage{
stock_plot(
  ticker1,
  ticker2,
  start = "2013-10-20",
  end = "2020-10-20",
  color1 = "dodgerblue3",
  color2 = "hotpink3"
)
}
\arguments{
\item{ticker1}{name of the first ticker to plot}

\item{ticker2}{name of the second ticker to plot}

\item{start}{start date, default set to "2013-10-20"}

\item{end}{end date, default set to "2020-10-20"}

\item{color1}{the color of the first plot, default set to "dodgerblue3"}

\item{color2}{the color of the first plot, default set to "hotpink3"}
}
\value{

}
\description{
stock_plot
}
\examples{
stock_plot(LIT,PCRFY,start="2013-10-20",end="2020-10-20",color1="dodgerblue3",color2="hotpink3")
}
