% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_cr.R
\name{t_cr}
\alias{t_cr}
\title{t_cr}
\usage{
t_cr(two_tail = TRUE, alpha, df)
}
\arguments{
\item{two_tail}{= TURE if it is a two-tail test, = FALSE if one-tail, default is two-tail}

\item{alpha}{the significance level of choice}

\item{df}{degree of freedom}
}
\value{

}
\description{
t_cr
}
\examples{
t_cr(two_tail=FALSE,0.05,5)
}
